requirements.txt
livekit-agents[openai,silero,elevenlabs]==1.0.11
python-dotenv==1.0.1
